enterprise_ai_backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
pydantic>=2.0
sqlalchemy>=2.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0
python-jose[cryptography]>=3.3
scikit-learn>=1.3
numpy>=1.24
httpx>=0.27
pytest>=8.0
# Postgres driver for prod - comment in when DATABASE_URL points at Postgres
# psycopg[binary]>=3.1
